Attempt to fix BeautiTest 2

+++ src/test/beast/app/BeautiTest.java
@@ -16,7 +16,8 @@ public class BeautiTest extends TestCase {
 	// test that beauti can merge an alignment with a template and write out a file
 	// this requires that the standard template can be read
 	public void testStandarBatchMode() {
-		try {
+		// ignore test if no X11 display available
+		if (! java.awt.GraphicsEnvironment.isHeadless()) {
 			File f = new File(sFile);
 			if (f.exists()) {
 				f.delete();
@@ -24,15 +25,14 @@ public class BeautiTest extends TestCase {
 			Beauti.main(("-template templates/HKY.xml -nex examples/nexus/dna.nex -out " + sFile + " -exitaction writexml").split(" "));
 			f = new File(sFile);
 			assertEquals(f.exists() && f.length() > 0, true);
-		} catch (HeadlessException e) {
-			// ignore test if no X11 display available
 		}
 	}
 
 	@Test
 	// as testStandarBatchMode() but for the *Beast template 
 	public void testStarBeastBatchMode() {
-		try {
+		// ignore test if no X11 display available
+		if (! java.awt.GraphicsEnvironment.isHeadless()) {
 			File f = new File(sFile);
 			if (f.exists()) {
 				f.delete();
@@ -40,8 +40,6 @@ public class BeautiTest extends TestCase {
 			Beauti.main(("-template templates/StarBeast.xml -nex examples/nexus/26.nex -nex examples/nexus/29.nex -out " + sFile + " -exitaction writexml").split(" "));
 			f = new File(sFile);
 			assertEquals(f.exists() && f.length() > 0, true);
-		} catch (HeadlessException e) {
-			// ignore test if no X11 display available
 		}
 	}
 }
